Feature Extractor
Added a test arff file: faces_50i-49d.arff
The pattern for the file name is this: 50i refers to 50 instances, 49d refers to 49 dimensions.

Finished the feature_extractor.py program.  Run python ./feature_extractor.py --help to see the help options.

Modified image_to_csv.py program to allow a .pic file to be used for image preferences.  This will allow us to go through each image file and choose whether we want to include it or leave it out of our final data set.  The .pic file keeps track of our preference so we don't have to repeat the process everytime.